Potential fix for code scanning alert no. 403: Wrong type of arguments to formatting function

+++ tests/testAtomic.cpp
@@ -54,7 +54,7 @@ struct TestState {
 	
 	bool Read() 
 	{
-		int val = fTable[0];
+		long val = fTable[0];
 		fReadCounter++;
 		for (int i = 0; i < SIZE; i++) {
 			if (fTable[i] != val) {
